Hook runInNewContext and runInContext too.
When --hook-run-in-context is passed, hook runInContext and
runInNewContext, in addition to the already hooked runInThisContext.
Since the new hooked functions use a context object as the global
object, we need to extract the coverage information from the context
global and merge it with the real global once the hooked function is
done executing.